Add invisible walls for enemies in caves
As requested by christopho: the walls are in the back and are in the
same model as in dungeon 1, meaning they block enemies and blocks. I
made a different commit because I extrapolated the fact this maps
also needs the walls, so you can accept the commit for dungeons and
not the one for cave.